% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_nearest_taxa.R
\name{enrich_search_nearest_taxa_phylo}
\alias{enrich_search_nearest_taxa_phylo}
\title{Enrich distance-to-SGB vector results}
\usage{
enrich_search_nearest_taxa_phylo(
  dist.vec,
  taxonomy = "mta",
  chocophlan_datestamp,
  tse = NULL
)
}
\arguments{
\item{dist.vec}{\code{double} vector of distances. the vector's names are the 
SGBs.}

\item{taxonomy}{\code{character} vector, either `mpa` (default) or `GTDB`.
NB: if you are interested in eukaryotes, use only `mpa`, GTDB lacks eukaryotes
taxonomy}

\item{chocophlan_datestamp}{A datestamp string or value identifying the
ChocoPhlAn database version (e.g. \code{"202307"}). Passed to
\code{validate_chocophlan_version()} before use.}

\item{tse}{\code{TreeSummarizedExperiment}. If passed, it is used to add one
extra column to the output saying whether that SGB is present in it.}
}
\value{
\code{data.frame} of full taxonomies and their distance to the 
 `SGB_target` provided in \code{search_nearest_taxa_phylo()}. if tse is provided,
 an extra column with presence in its rowData$SGB is provided (TRUE/FALSE).
}
\description{
After running `search_nearest_taxa_phylo()`, the output can be passed
to this function to format is as complete taxonomy data.frame.
}
\examples{
library(ChocoPhlAnTaxonomies)
suppressMessages(library(TreeSummarizedExperiment))
data(JoS_2022.tse)
JoS_2022.tse <- AddPhyloTree_to_mpa_tse(JoS_2022.tse, chocophlan_datestamp = "202403")

#############################################################################
# example with rowTree of the tse, all will be present   ####################

dist.vec <- search_nearest_taxa_phylo("t__SGB12137", rowTree(JoS_2022.tse))

## example without tse as argument
head(enrich_search_nearest_taxa_phylo(dist.vec, taxonomy = "GTDB", 
 chocophlan_datestamp = "202403"))

## example with tse as argument
head(enrich_search_nearest_taxa_phylo(dist.vec, taxonomy = "GTDB", 
chocophlan_datestamp = "202403", tse = JoS_2022.tse))

#############################################################################
# example with full phylogenetic tree, most won't be present ################
phyloTree.list <- load_mpa_tree("202403")
dist.vec <- search_nearest_taxa_phylo("t__SGB12137", phyloTree.list$Phylo)

## example without tse as argument
head(enrich_search_nearest_taxa_phylo(dist.vec, taxonomy = "GTDB", 
   chocophlan_datestamp = "202403"))

## example with tse as argument
head(enrich_search_nearest_taxa_phylo(dist.vec, taxonomy = "GTDB", 
   chocophlan_datestamp = "202403", tse = JoS_2022.tse))

}
\seealso{
\code{\link{search_nearest_taxa_phylo}}, \code{\link{search_nearest_taxa_phylo}}
}
