% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_nearest_taxa.R
\name{search_nearest_taxa_phylo}
\alias{search_nearest_taxa_phylo}
\title{Find phylogenetically closest taxa to a target}
\usage{
search_nearest_taxa_phylo(SGB_target, phyloTree)
}
\arguments{
\item{SGB_target}{\code{character} with SGB name with or without the "t__" 
prefix.}

\item{phyloTree}{\code{phylo} class tree. Ideally the user should consider 
loading trees with the package on GitHub 'g-antonello/ChocoPhlAnTaxonomies'.
This ensures that tip labels are identical to taxonomies' terminal ID 
(`Species` for chocophlan 2019, `SGB` for all later releases)}
}
\value{
\code{double} vector with all taxa distances sorted in increasing
order. The vector is named after the SGB considered. All distances are 
relative to the \code{SGB_target} from the \code{phyloTree} object. The 
self-distance, always 0, is removed from the vector. The vector returned 
therefore has 1 element less than the number of tips in the \code{phyloTree}.
}
\description{
This function ranks SGBs in relation to phylogenetic similarity to a target
SGB. Important: the function does not check whether the target SGB is present
in a dataset: the user should be mindful of what they are searching.
}
\examples{
library(ChocoPhlAnTaxonomies)
suppressMessages(library(TreeSummarizedExperiment))
phyloTree.list <- load_mpa_tree("202403")
dist.vec <- search_nearest_taxa_phylo("t__SGB12137", phyloTree.list$Phylo)

SGBs_distances <- search_nearest_taxa_phylo("t__SGB12137", phyloTree.list$Phylo)
head(SGBs_distances)
}
