% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WallenZD_2022.tse}
\alias{WallenZD_2022.tse}
\title{Wallen ZD 2022 MetaPhlAn 3.0 relative abundance example data}
\format{
##
   class: TreeSummarizedExperiment 
   dim: 719 724 
   metadata(0):
     assays(1): metaphlan
   rownames(719): s__Methanobacterium_sp_MB1 s__Methanobrevibacter_arboriphilus ...
   s__Saccharomyces_cerevisiae s__Blastocystis_sp_subtype_1
   rowData names(8): kingdom phylum ... species clade_name
   colnames(724): DP310 DP309 ... DC169 DC168
   colData names(57): sample_name Case_status ... collection_method total_sequences
   reducedDimNames(0):
     mainExpName: NULL
   altExpNames(6): kingdom phylum ... family genus
   rowLinks: NULL
   rowTree: NULL
   colLinks: NULL
   colTree: NULL
}
\source{
<https://zenodo.org/records/7246185>
}
\usage{
WallenZD_2022.tse
}
\description{
Relative abundance, taxonomy and sample metadata (colData) in 
`TreeSummarizedExperiment` format. The data were processed with MetaPhlAn
3.0, using the 201901 database. rows are taxa, columns are samples. 
Paper: https://doi.org/10.1038/s41467-022-34667-x
}
\keyword{datasets}
